+++ lua/mini/base16.lua
@@ -154,15 +154,6 @@ local H = {}
 ---@usage `require('mini.base16').setup({})` (replace `{}` with your `config`
 ---   table; `config.palette` should be a table with colors)
 MiniBase16.setup = function(config)
-  -- TODO: Remove after Neovim<=0.6 support is dropped
-  if vim.fn.has('nvim-0.7') == 0 then
-    vim.notify(
-      '(mini.base16) Neovim<0.7 is soft deprecated (module works but not supported).'
-        .. ' It will be deprecated after Neovim 0.9.0 release (module will not work).'
-        .. ' Please update your Neovim version.'
-    )
-  end
-
   -- Export module
   _G.MiniBase16 = MiniBase16